refactor(source/bucket): split bucket.go into endpoint.go + transport.go + traversal.go
Mirror to the earlier git source split (#277). Three test files
(endpoint_test.go, transport_test.go, traversal_test.go) named what
their production-side counterparts should be, but the impls lived
in bucket.go alongside the Fetch entry, walkBucket, downloadObject,
and resolveCredentials.

- endpoint.go (new, 39 lines): normalizeEndpoint + schemeFor.
  Covered by endpoint_test.go's 2 tests.
- transport.go (new, 49 lines): resolveTransport. Covered by
  transport_test.go's 6 tests.
- traversal.go (new, 30 lines): safeJoinUnderSlot. Covered by
  traversal_test.go's TestSafeJoinUnderSlot.
- bucket.go (293 → 213 lines): Fetcher + Fetch + walkBucket +
  downloadObject + resolveCredentials.

resolveCredentials stays in bucket.go for now — no
credentials_test.go counterpart calls for it. Behavior unchanged;
full suite + lint green.
